Added byronsanchez's fork of Jekyll as a submodule.
The byronsanchez fork of Jekyll has set "safe" to "false" in the
converters and generators who had it turned on. When safe is true, the
converters and generators cannot be overwritten via the Jekyll _plugins/
directory. However, when safe is false, these files can be overwritten.

hackBytes.com overwrites these files to provide a [gruesome] bugfix to
pygments.rb on the Jekyll layer. This makes sure syntax highlighting
can display line numbers properly. In addition, the overwrites allow
hackbytes.com to extend the syntax highlighting functionality so users
can define the language using standard markdown-style codeblocks. The
language can even be defined as plaintext, to invoke a new style that is
easier on the eyes for reading normal text.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>